Moved SearchBar.js into its own 'SearchBar' folder within /components/; created styling for SearchBar to mimic wireframe SearchBar; made button invisible and added event listener for "enter" key for search submission (again to mimic wireframe)
