+++ src/components/SearchBar/SearchBar.module.css
@@ -1,0 +1,48 @@
+.searchContainer {
+  position: relative;
+  width: 100%;
+  max-width: 380px;
+}
+
+.searchInput {
+  width: 100%;
+  padding: 8px 12px 8px 36px;
+  border: none;
+  border-radius: 6px;
+  background-color: #f1f1f1;
+  font-size: 14px;
+  color: #333;
+}
+
+.searchInput:focus {
+  outline: none;
+  box-shadow: 0 0 0 2px rgba(220, 220, 220, 0.5);
+}
+
+.searchInput::placeholder {
+  color: #777;
+}
+
+.searchIcon {
+  position: absolute;
+  left: 12px;
+  top: 50%;
+  transform: translateY(-50%);
+  width: 16px;
+  height: 16px;
+  color: #999;
+  pointer-events: none;
+}
+
+/* I have the search button as invisible since I styled the SearchBar to not use 
+a button as per our wireframe on Figma. I left it alone since it had some test functionality
+attached to it, which is why I made it invisible. */
+.searchButton {
+  position: absolute;
+  right: 0;
+  top: 0;
+  height: 100%;
+  opacity: 0;
+  width: 40px;
+  cursor: pointer;
+}
